tests: advance the rail clock and isolate design checks
