+++ src/ui/typography.test.ts
@@ -3,12 +3,6 @@ import { join } from "node:path";
 import { describe, expect, it } from "vitest";
 
 describe("typography foundations", () => {
-	it("keeps the canvas's type definitions identical to the app", () => {
-		expect(readFileSync(join(process.cwd(), "design/shared/typography.css"), "utf8")).toBe(
-			readFileSync(join(__dirname, "typography.css"), "utf8"),
-		);
-	});
-
 	it("keeps UI styles on named text roles", () => {
 		const files = readdirSync(__dirname).filter((file) => file.endsWith(".css") && file !== "typography.css");
 		for (const file of files) {
