+++ vitest.config.ts
@@ -24,23 +24,9 @@ export default defineConfig({
 		alias: pins,
 	},
 	test: {
-		include: ["src/**/*.test.ts"],
-		/**
-		 * One retry on CI, and none here.
-		 *
-		 * This suite spends most of its time waiting on things it started — daemons,
-		 * headless browsers, worker processes, filesystem watchers — and it runs 150 files
-		 * at once to stay under fifteen minutes. Those two facts together mean a test can
-		 * lose its slot long enough to run past a wait, and which test that happens to is
-		 * scheduling noise: a full run flakes roughly one time in three, on a different
-		 * test each time, every one of which passes on its own. A retry answers exactly
-		 * that and nothing else, because a test that is actually broken fails twice.
-		 *
-		 * It is off locally on purpose. A flake in front of a person is information, and
-		 * this is where it should be paid attention to rather than smoothed over; the
-		 * tight budget it exposes is worth fixing at the source, the way `reachForPill`
-		 * and the machine-state waits were.
-		 */
+		include: ["src/**/*.test.ts", ".github/scripts/*.test.ts"],
+		setupFiles: ["./src/test-setup.ts"],
+		// Keep one CI retry for process/browser scheduling noise; local failures remain visible.
 		retry: process.env.CI === undefined ? 0 : 1,
 	},
 });
